Refactor imports to use hooks instead of theme provider
