+++ projects/js-packages/charts/src/components/line-chart/stories/index.stories.tsx
@@ -1,4 +1,4 @@
-import { useChartTheme } from '../../../providers/theme';
+import { useChartTheme } from '../../../hooks';
 import { legendArgTypes } from '../../../stories/legend-config';
 import LineChart from '../line-chart';
 import { lineChartStoryArgs, lineChartMetaArgs } from './config';
